+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## [Unreleased]
 
+## [0.8.0] - 2026-07-22
+
 ### Added
 
 - Added a toggle in VPN settings to restrict the WireGuard tunnel to urs's own traffic only.